// packages/books-service/src/types/api.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/categories": {
    get: operations["categories"];
  };
  "/books/{id}": {
    get: operations["book"];
  };
  "/books": {
    get: operations["books"];
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    Category: {
      id: string;
      name: string;
    };
    Book: {
      id: string;
      title: string;
      authorId: string;
      categoryId: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {

  categories: {
    responses: {
      200: {
        content: {
          "application/json": components["schemas"]["Category"][];
        };
      };
    };
  };
  book: {
    parameters: {
      path: {
        id: string;
      };
    };
    responses: {
      200: {
        content: {
          "application/json": components["schemas"]["Book"];
        };
      };
    };
  };
  books: {
    parameters: {
      query: {
        ids: string[];
      };
    };
    responses: {
      200: {
        content: {
          "application/json": components["schemas"]["Book"][];
        };
      };
    };
  };
}